+++ scripts/esbuild.js
@@ -5,6 +5,7 @@ const fs = require('fs')
 
 const repoDir = path.dirname(__dirname)
 const npmDir = path.join(repoDir, 'npm', 'esbuild')
+const version = require(path.join(npmDir, 'package.json')).version
 
 function buildNativeLib(esbuildPath) {
   const libDir = path.join(npmDir, 'lib')
@@ -18,6 +19,7 @@ function buildNativeLib(esbuildPath) {
     path.join(repoDir, 'lib', 'install.ts'),
     '--outfile=' + path.join(npmDir, 'install.js'),
     '--target=es2015',
+    '--define:ESBUILD_VERSION=' + JSON.stringify(version),
     '--platform=node',
   ], { cwd: repoDir })
 
@@ -29,6 +31,7 @@ function buildNativeLib(esbuildPath) {
     '--target=es2015',
     '--format=cjs',
     '--define:WASM=false',
+    '--define:ESBUILD_VERSION=' + JSON.stringify(version),
     '--platform=node',
   ], { cwd: repoDir })
 
@@ -53,6 +56,7 @@ function buildWasmLib(esbuildPath) {
     '--target=es2015',
     '--format=cjs',
     '--define:WASM=true',
+    '--define:ESBUILD_VERSION=' + JSON.stringify(version),
     '--platform=node',
   ], { cwd: repoDir })
 
@@ -75,6 +79,7 @@ function buildWasmLib(esbuildPath) {
   const workerMinCode = childProcess.execFileSync(esbuildPath, [
     path.join(repoDir, 'lib', 'worker.ts'),
     '--minify',
+    '--define:ESBUILD_VERSION=' + JSON.stringify(version),
   ], { cwd: repoDir }).toString().trim()
 
   // Generate "npm/esbuild-wasm/browser.js"
@@ -86,6 +91,7 @@ function buildWasmLib(esbuildPath) {
     '--target=es2015',
     '--minify',
     '--format=cjs',
+    '--define:ESBUILD_VERSION=' + JSON.stringify(version),
     '--define:WEB_WORKER_SOURCE_CODE=' + JSON.stringify(wasmExecMinCode + workerMinCode),
   ], { cwd: repoDir }).toString()
   fs.writeFileSync(path.join(libDir, 'browser.js'), umdPrefix + browserJs.trim() + umdSuffix)
@@ -107,7 +113,6 @@ exports.installForTests = dir => {
 
   // Install the "esbuild" package
   const env = { ...process.env, ESBUILD_BIN_PATH_FOR_TESTS: esbuildPath }
-  const version = require(path.join(npmDir, 'package.json')).version
   fs.writeFileSync(path.join(dir, 'package.json'), '{}')
   childProcess.execSync(`npm pack --silent "${npmDir}"`, { cwd: dir, stdio: 'inherit' })
   childProcess.execSync(`npm install --silent --no-audit --progress=false esbuild-${version}.tgz`, { cwd: dir, env, stdio: 'inherit' })